+++ scripts/test_archive.sh
@@ -128,6 +128,8 @@ function run_parquet_to_auditor_db() {
   check_if_records_are_restored_by_rust_script
   
   fill_auditor_db_group4
+  
+  rm -rf ../../../../archived_records
 
   sleep 25